requirements.txt
aiofiles >= 0.5.0
attrs==20.2.0
bcrypt==3.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
configparser==5.0.1
coverage==5.3
cryptography==3.1.1
docopt==0.6.2
google-api-core==1.23.0
google-auth==1.22.1
google-cloud-core==1.4.3
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
hdfs==2.5.8
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
minio==6.0.0
numpy==1.19.2
packaging==20.4
pandas==1.1.3
paramiko==2.7.2
pluggy==0.13.1
protobuf==3.13.0
py==1.9.0
py-cpuinfo==7.0.0
pyarrow >= 0.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
pyorc==0.4.0
pyparsing==2.4.7
pytest==6.1.1
pytest-benchmark==3.2.3
pytest-cov==2.10.1
pytest-timeout==1.4.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
toml==0.10.1
urllib3==1.25.11
zipp==3.3.1
keyring==21.5.0
hvac==0.10.5